feat(credentials): séparation en plusieurs composants
