+++ front/src/components/Login.jsx
@@ -125,7 +125,7 @@ export default function Login() {
               autoComplete="username"
             />
             <Field
-              label={t('credentials.password.placeholder')}
+              label={t('credentials.password.label')}
               name="password"
               hasError={error !== ''}
               required={true}
