+++ front/src/components/button.module.scss
@@ -95,7 +95,7 @@ a.icon:not(.primary) {
   color: currentColor;
   cursor: pointer;
   font-size: inherit;
-  text-decoration: underline;
+  text-decoration: none;
 
   &:hover,
   &:focus {
