+++ front/src/locales/en/translation.json
@@ -1,5 +1,5 @@
 {
-  "user.account.title": "Account information",
+  "user.account.title": "User profile",
   "user.account.displayName": "Display name",
   "user.account.firstName": "First name",
   "user.account.lastName": "Last name",
@@ -40,7 +40,7 @@
   "credentials.authentication.error.unknownService": "Unknown authentication provider.",
   "credentials.authentication.error.title": "Authentication failure",
   "credentials.authentication.back": "Back to Stylo",
-  "credentials.authentication.linkedService.description": "Stylo account linked to {{ service }} ({{ token }}).",
+  "credentials.authentication.linkedService.description": "Stylo account linked to {{ service }}.",
   "credentials.authentication.unlinkedService.description": "No linked account.",
   "credentials.authentication.linkButton": "Link",
   "credentials.authentication.linkLabel": "Link your Stylo account to {{ service }} service.",
@@ -193,12 +193,11 @@
   "workspace.createForm.buttonText": "Create",
   "contact.userNotFound.message": "No user found for this email adress",
   "contact.searchField.placeholder": "Contact email",
-  "credentials.changePassword.title": "Change password",
-  "credentials.changePassword.para": "This section is strictly private, changing your password will only affect your combination of username/email and password.",
-  "credentials.oldPassword.placeholder": "Old password",
-  "credentials.password.placeholder": "Password",
-  "credentials.newPassword.placeholder": "New password",
-  "credentials.confirmNewPassword.placeholder": "Confirm new password",
+  "credentials.changePassword.title": "Password",
+  "credentials.oldPassword.label": "Old password",
+  "credentials.password.label": "Password",
+  "credentials.newPassword.label": "New password",
+  "credentials.confirmNewPassword.label": "Confirm new password",
   "export.format.label": "Formats",
   "export.bibliography.label": "Bibliography style",
   "export.additionnalOptions.label": "Other settings",
