+++ front/src/locales/es/translation.json
@@ -127,11 +127,10 @@
   "workspace.createForm.buttonText": "Crear",
   "contact.userNotFound.message": "No hay usuarios en esta dirección de correo electrónico",
   "contact.searchField.placeholder": "Correo electrónico de contacto",
-  "credentials.changePassword.title": "Cambiar contraseña",
-  "credentials.changePassword.para": "Esta sección es estrictamente privada. El cambio de contraseña sólo afectará a su combinación de nombre de usuario/correo electrónico y contraseña.",
-  "credentials.oldPassword.placeholder": "Contraseña antigua",
-  "credentials.newPassword.placeholder": "Nueva contraseña",
-  "credentials.confirmNewPassword.placeholder": "Confirmar nueva contraseña",
+  "credentials.changePassword.title": "Contraseña",
+  "credentials.oldPassword.label": "Contraseña antigua",
+  "credentials.newPassword.label": "Nueva contraseña",
+  "credentials.confirmNewPassword.label": "Confirmar nueva contraseña",
   "export.format.label": "Formatos",
   "export.sectionChapters.numbered": "Sección y capítulos: numerados",
   "export.sectionChapters.unnumbered": "Sección y capítulos: sin numerar",
@@ -274,7 +273,7 @@
   "article.metadata.production.entities.media.all": " ",
   "article.metadata.people.forename": "Nombre",
   "form.addPerson.title": "Añadir una persona",
-  "credentials.password.placeholder": "Contraseña",
+  "credentials.password.label": "Contraseña",
   "credentials.updatePassword.confirmButton": "Actualizar",
   "export.nocite.all": "Todas las referencias",
   "article.metadata.journal.email": "Correo electrónico",
@@ -304,7 +303,7 @@
   "article.metadata.language.addItem": "Agregar un idioma",
   "export.nocite.label": "Referencias",
   "header.languages": "Cambiar el idioma de la interfaz",
-  "user.account.title": "Información de la cuenta",
+  "user.account.title": "Perfil de usuario",
   "export.bibliography.label": "Estilo de bibliografía",
   "article.metadata.senspublic.categories": "Categorías",
   "article.metadata.titleAndAbstract": "Título y resumen",
