+++ front/src/locales/fr/translation.json
@@ -1,5 +1,5 @@
 {
-  "user.account.title": "Information du compte",
+  "user.account.title": "Profil utilisateur",
   "user.account.displayName": "Nom d'affichage",
   "user.account.firstName": "Prénom",
   "user.account.lastName": "Nom",
@@ -14,6 +14,8 @@
   "user.account.copyApiKey": "Copier la clé d'accès à l'API dans le presse-papier",
   "user.account.apiKeyValue": "La valeur de la clé d'API est : {{token}}",
   "user.languages.title": "Langues",
+  "user.authProviders.title": "Comptes externes",
+  "user.authProviders.description": "Relier des comptes externes permet de bénéficier de fonctionnalités avancées (synchronisation de données, authentification, etc.).",
   "credentials.manage": "Gérer mon compte",
   "credentials.register.title": "Créer un compte Stylo local",
   "credentials.register.requiredFields": "Champs obligatoires",
@@ -39,7 +41,7 @@
   "credentials.authentication.error.unknownService": "Service d'authentification inconnu.",
   "credentials.authentication.error.title": "Échec de l'authentification",
   "credentials.authentication.back": "Revenir à Stylo",
-  "credentials.authentication.linkedService.description": "Compte Stylo connecté à {{ service }} ({{ token }}).",
+  "credentials.authentication.linkedService.description": "Compte Stylo connecté à {{ service }}.",
   "credentials.authentication.unlinkedService.description": "Pas de compte connecté.",
   "credentials.authentication.linkButton": "Connecter",
   "credentials.authentication.linkLabel": "Connecter votre compte Stylo au service {{ service }}.",
@@ -193,12 +195,11 @@
   "workspace.createForm.buttonText": "Créer",
   "contact.userNotFound.message": "Aucun utilisateur à cette adresse courriel",
   "contact.searchField.placeholder": "Email du contact",
-  "credentials.changePassword.title": "Changer le mot de passe",
-  "credentials.changePassword.para": "Cette section est strictement privée, la modification de votre mot de passe n'affectera que votre combinaison nom d'utilisateur/email et mot de passe.",
-  "credentials.oldPassword.placeholder": "Ancien mot de passe",
-  "credentials.password.placeholder": "Mot de passe",
-  "credentials.newPassword.placeholder": "Nouveau mot de passe",
-  "credentials.confirmNewPassword.placeholder": "Confirmer le nouveau mot de passe",
+  "credentials.changePassword.title": "Mot de passe",
+  "credentials.oldPassword.label": "Ancien mot de passe",
+  "credentials.password.label": "Mot de passe",
+  "credentials.newPassword.label": "Nouveau mot de passe",
+  "credentials.confirmNewPassword.label": "Confirmer le nouveau mot de passe",
   "export.format.label": "Formats",
   "export.bibliography.label": "Style bibliographique",
   "export.toc.label": "Table des matières",
